+++ src/Falcata.BillPlanner.Application.Interfaces/Repositories/IAccountCommandRepository.cs
@@ -3,8 +3,8 @@ using Falcata.BillPlanner.Domain.Models.BillPlanner.Accounts;
 
 namespace Falcata.BillPlanner.Application.Interfaces.Repositories;
 
-public interface IAccountCommandRepository: ICreateCommandRepository<Account, (long AccountId, int AccountTypeId)>,
-    IUpdateCommandRepository<Account, (long AccountId, int AccountTypeId)>
+public interface IAccountCommandRepository: ICreateCommandRepository<Account, (int AccountId, int AccountTypeId)>,
+    IUpdateCommandRepository<Account, (int AccountId, int AccountTypeId)>
 {
     
 }
